2nd. Don't work logLevel FINE. Why?

+++ src/main/java/org/example/utils/StatisticsGeneratorNew.java
@@ -54,7 +54,7 @@ public class StatisticsGeneratorNew {
                     .average();
             statistics.setAvgExamScore(0);
             avgExamScore.ifPresent(value -> statistics.setAvgExamScore(
-                    BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
+                    (float)BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
 //                    (float) BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP).doubleValue()));
         });
